version 1.0 release 3/2/2020
The app has been finished completely and is now ready for version 1.0 upload. Version 1.1 will allow users to vote on who they want to be elected.

This update finishes Weld's modals, as well as the events tab.

Next update will include a 3rd tab which is where the voting will occur.